pyproject.toml
[tool.poetry]
name = "drf_test"
version = "0.1.0"
description = ""
authors = ["naga3 <osanaga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
flake8 = "^3.7"
autopep8 = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
